app/requirements.txt
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
boto3==1.21.23
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
fastapi==0.72.0
filelock==3.4.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.6.1
ipython==8.0.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
lxml==4.7.1
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.1
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pure-eval==0.2.1
py==1.11.0
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.0.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
seaborn==0.11.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
stack-data==0.1.3
starlette==0.17.1
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
wcwidth==0.2.5
Werkzeug==2.0.3
mlxtend==0.19.0
